Updated BroadcastLoginSignal to include IP Address and Port Implemented client to server BroadcastLogin  - Still need to handle incoming login broadcasts on client side Implemented server ability to collect a list of visible contacts for a specified user  - Still need to send the data to the client Possibly need to re-factor tables to include Nickname, Personal Message and Status at a later time
git-svn-id: file:///e/SVN/serious_messenger@44 b1debafe-a605-4e1b-80f9-83a58a4e6cd1

+++ Serious Client/src/client/Contact.java
@@ -24,11 +24,11 @@ public class Contact extends ContactData {
 	
 	private Client m_client;
 	
-	public Contact(String userName, String nickName, String personalMessage, int status, Font font, Color textColour, String ipAddress, int port) {
+	public Contact(String userName, String nickName, String personalMessage, byte status, Font font, Color textColour, String ipAddress, int port) {
 		this(userName, nickName, personalMessage, status, font, textColour, parseIPAddress(ipAddress), port);
 	}
 	
-	public Contact(String userName, String nickName, String personalMessage, int status, Font font, Color textColour, InetAddress ipAddress, int port) {
+	public Contact(String userName, String nickName, String personalMessage, byte status, Font font, Color textColour, InetAddress ipAddress, int port) {
 		super(userName, nickName, personalMessage, status, font, textColour);
 		setIPAddress(ipAddress);
 		setPort(port);
